requirements.txt
# Core Framework Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
python-multipart==0.0.6

# AI/ML Dependencies (Updated for Python 3.13 compatibility)
torch>=2.5.0
torchvision>=0.20.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Database Dependencies
asyncpg>=0.29.0
redis>=5.0.0
alembic>=1.12.0
SQLAlchemy>=2.0.20

# Video Processing (Simplified for Codespaces)
# Note: Full DeepStream SDK requires NVIDIA GPU setup
# Using CPU-only alternatives for testing

# Computer Vision
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
click>=8.1.0
rich>=13.7.0
httpx>=0.25.0
aiofiles>=23.2.0

# Monitoring and Logging
prometheus-client>=0.19.0
structlog>=23.2.0

# Development Dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.11.0
flake8>=6.1.0

# Production Dependencies
gunicorn>=21.2.0